+++ api/accounts/user/profile/api_views.py
@@ -1,7 +1,7 @@
 from django.db import connection
 
 from api.api_views import APIView
-from api.signals import user_relationship_changed
+from api.signals import user_profile_changed
 from api.exceptions import OperationNotSupported
 from api.accounts.user.profile.serializers import UserProfileSerializer
 from api.accounts.user.utils import get_user, get_user_profiles
@@ -44,9 +44,12 @@ class UserProfileView(APIView):
             return FailureTaskResponse(self.request, ser.errors, obj=profile, dc_bound=False)
 
         ser.save()
-        connection.on_commit(lambda: user_relationship_changed.send(user_name=ser.object.user.username))
-        return SuccessTaskResponse(self.request, ser.data, obj=self.user, detail_dict=ser.detail_dict(),
-                                   owner=ser.object.user, msg=LOG_PROFILE_UPDATE, dc_bound=False)
+        res = SuccessTaskResponse(self.request, ser.data, obj=self.user, detail_dict=ser.detail_dict(),
+                                  owner=ser.object.user, msg=LOG_PROFILE_UPDATE, dc_bound=False)
+        task_id = res.data.get('task_id')
+        connection.on_commit(lambda: user_profile_changed.send(task_id, user_name=self.user.username))  # Signal!
+
+        return res
 
     # noinspection PyMethodMayBeStatic
     def delete(self):