+++ api/dc/base/serializers.py
@@ -275,10 +275,13 @@ class DcSettingsSerializer(s.InstanceSerializer):
     VMS_ZONE_ENABLED = s.BooleanField(label='VMS_ZONE_ENABLED',  # Module
                                       help_text=_('Whether to enable support for SunOS and Linux zones in '
                                                   'this virtual datacenter.'))
-
     VMS_VM_DEFINE_LIMIT = s.IntegerField(label='VMS_VM_DEFINE_LIMIT', required=False,
                                          help_text=_('Maximum number of virtual servers that can be defined in '
                                                      'this virtual datacenter.'))
+    VMS_VM_CPU_CAP_REQUIRED = s.BooleanField(label='VMS_VM_CPU_CAP_REQUIRED',
+                                             help_text='When disabled, the vCPUs server parameter on SunOS and LX Zones'
+                                                       ' can be set to 0, which removes the compute node CPU limit'
+                                                       ' (cpu_cap) for the virtual server.')
     VMS_VM_STOP_TIMEOUT_DEFAULT = s.IntegerField(label='VMS_VM_STOP_TIMEOUT_DEFAULT',
                                                  help_text='Default time period (in seconds) for a graceful VM stop or '
                                                            'reboot, after which a force stop/reboot is send to the VM '
@@ -687,7 +690,7 @@ class DefaultDcSettingsSerializer(DcSettingsSerializer):
                                                 help_text=_('Existing Zabbix host group, which will be used for all '
                                                             'monitored compute nodes.'))
     MON_ZABBIX_HOSTGROUPS_NODE = s.ArrayField(label='MON_ZABBIX_HOSTGROUPS_NODE', max_items=32, required=False,
-                                              help_text=_('List of other existing Zabbix host groups, which will be '
+                                              help_text=_('List of other Zabbix host groups, which will be '
                                                           'used for all monitored compute nodes.'))
     MON_ZABBIX_TEMPLATES_NODE = s.ArrayField(label='MON_ZABBIX_TEMPLATES_NODE', max_items=128, required=False,
                                              help_text=_('List of existing Zabbix templates, which will be used for all'