+++ api/mon/backends/__init__.py
@@ -1,10 +1,11 @@
 from __future__ import absolute_import
+
 from django.conf import settings
 
-from . import zabbix
-from . import dummy
+from api.mon.backends import zabbix
+from api.mon.backends import dummy
 
-__all__ = ('get_monitoring', 'del_monitoring', 'MonitoringBackend')
+__all__ = ('get_monitoring', 'del_monitoring', 'MonitoringBackend', 'MonitoringServer')
 
 BACKEND_ALIASES = {
     'dummy': dummy,
@@ -16,6 +17,10 @@ DEFAULT_BACKEND = 'zabbix'
 backend = BACKEND_ALIASES[getattr(settings, 'MONITORING_BACKEND', DEFAULT_BACKEND)]
 
 MonitoringBackend = backend.MonitoringBackendClass
+
+MonitoringServer = backend.MonitoringServerClass
+
+MonitoringBackend.server_class = MonitoringServer
 
 
 def get_monitoring(dc, **kwargs):