+++ api/mon/tasks.py
@@ -6,3 +6,5 @@ from api.mon.vm.tasks import *  # noqa: F401,F403
 from api.mon.node.tasks import *  # noqa: F401,F403
 # noinspection PyUnresolvedReferences
 from api.mon.alerting.tasks import *  # noqa: F401,F403
+# noinspection PyUnresolvedReferences
+from api.mon.alerting.action.tasks import *  # noqa: F401,F403